+++ src/asp/Camera/SatSim.cc
@@ -21,6 +21,7 @@
 #include <asp/Core/SatSimBase.h>
 #include <asp/Core/CameraTransforms.h>
 #include <asp/Core/Common.h>
+#include <asp/Core/DemUtils.h>
 #include <asp/Camera/SatSim.h>
 #include <asp/Camera/CsmModel.h>
 #include <asp/Camera/SyntheticLinescan.h>
@@ -44,34 +45,6 @@ using namespace vw::geometry;
 namespace fs = boost::filesystem;
 
 namespace asp {
-
-// Find a handful of valid DEM values and average them. It helps later when
-// intersecting with the DEM, especially for Mars, where the DEM heights ca be
-// very far from the datum. 
-double findDemHeightGuess(vw::ImageViewRef<vw::PixelMask<float>> const& dem) {
-
-  double height_guess = 0.0;
-  bool found = false;
-  double sum = 0.0, num = 0.0;
-  for (double row = 0; row < dem.rows(); row += dem.rows()/10.0) {
-    for (double col = 0; col < dem.cols(); col += dem.cols()/10.0) {
-      if (is_valid(dem(col, row))) {
-        sum += dem(col, row).child();
-        num++;
-        if (num > 20) {
-          // Those are enough, as going on for too long may take too much time
-          found = true;
-          break;
-        }
-      }
-    }
-    if (found) break;
-  }
-  if (num > 0) 
-    height_guess = sum/num;
-    
-  return height_guess;
-} // End function findDemHeightGuess()
 
 // Compute satellite position, and along / across track normalized vectors in
 // ECEF coordinates, given the first and last proj points and a value t giving
@@ -1445,13 +1418,13 @@ void genImages(SatSimOptions const& opt,
     bool has_nodata = true;
     SatSimOptions local_opt = opt;
     local_opt.raster_tile_size = vw::Vector2i(512, 512);
-
     block_write_gdal_image(image_names[i], 
-      vw::apply_mask(SynImageView(opt, cams[i], 
-      dem_georef, dem, height_guess, ortho_georef, ortho, ortho_nodata_val), ortho_nodata_val),
-      has_georef, ortho_georef,  // ortho georef will not be used, but needed for the api
-      has_nodata, ortho_nodata_val, // borrow the nodata from ortho
-      local_opt, vw::TerminalProgressCallback("", "\t--> "));
+                           vw::apply_mask(SynImageView(opt, cams[i], dem_georef, dem, 
+                                                       height_guess, ortho_georef, 
+                                                       ortho, ortho_nodata_val), 
+                                          ortho_nodata_val),
+                           has_georef, ortho_georef, has_nodata, ortho_nodata_val, 
+                           local_opt, vw::TerminalProgressCallback("", "\t--> "));
   }  
 
   // Write the list of images only if we are not skipping the first camera